refactor: move CurrencyTypes to application
