+++ src/componentLibrary/CoinIcon/CoinIcon.stories.tsx
@@ -1,6 +1,6 @@
 import { Story, Meta } from '@storybook/react';
 
-import { CurrencySymbols } from '../';
+import { CurrencySymbols } from 'types/currency';
 import { CoinIcon, CoinIconProps } from '.';
 
 const Template: Story<CoinIconProps> = args => <CoinIcon {...args} />;
