+++ src/componentLibrary/VaultCard/VaultCard.stories.tsx
@@ -3,7 +3,7 @@ import {
   Meta
 } from '@storybook/react';
 
-import { CurrencySymbols } from '../';
+import { CurrencySymbols } from 'types/currency';
 import { VaultCard, VaultCardProps } from './';
 
 const Template: Story<VaultCardProps> = args => <VaultCard {...args} />;
