--- src/componentLibrary/types.ts
@@ -1,10 +1,0 @@
-enum CurrencySymbols {
-  DOT = 'DOT',
-  INTERBTC = 'interBTC',
-  INTR = 'INTR',
-  KBTC = 'KBTC',
-  KINT = 'KINT',
-  KSM = 'KSM',
-}
-
-export { CurrencySymbols };
