+++ src/utils/helpers/currency-pairs.ts
@@ -1,7 +1,6 @@
 import { CurrencyIdLiteral } from '@interlay/interbtc-api';
 
-import { CurrencySymbols } from 'componentLibrary';
-import { CurrencyPair, CurrencyPairs } from 'types/currency';
+import { CurrencyPair, CurrencyPairs, CurrencySymbols } from 'types/currency';
 
 const CURRENCY_PAIRS: CurrencyPairs = [
   {
